fix: correct FK column type mismatches across 7 entities
Several foreign-key columns were declared with a SQL type that
doesn't match the column they reference: numeric referencing an
integer primary key, or vice versa (e.g. Order.shipping_type_id was
'numeric' but ShippingType.type_id is 'int'). Postgres rejects
creating a foreign key constraint between incompatible types, so this
would have failed the very first time anyone ran `synchronize: true`
against a genuinely empty database - it only ever "worked" because
the live schema was created once and the entities drifted from it
afterward.

This was found by actually running the phase3 initial migration
against a real (WASM-compiled) Postgres via @electric-sql/pglite,
since there's no live Postgres available in this environment to
generate/verify a migration against otherwise - the first attempt
failed with exactly this error on the orders -> shipping_types FK,
and auditing the rest of the entities turned up 7 more instances of
the same bug.

Fixed by converging on `integer` for all of these ID columns (dropping
`numeric`, which was never appropriate for a plain identifier and had
the side effect of the pg driver returning it as a string instead of a
number):
- order_statuses.status_id, product_categories.category_id,
  target_audience.audience_id, skin_type.skin_type_id,
  product_type.product_type_id (their own primary keys)
- orders.status_id, orders.shipping_type_id
- order_items.order_id, order_items.product_id

+++ server/src/entities/orderStatus.entity.ts
@@ -4,7 +4,7 @@ import { OrderTracking } from './orderTracking.entity';
 
 @Entity({ schema: 'skinlab', name: 'order_statuses' })
 export class OrderStatus {
-  @PrimaryColumn({ type: 'numeric' })
+  @PrimaryColumn({ type: 'integer' })
   status_id: number;
 
   @Column({ type: 'text', unique: true })
